% Generated by roxygen2 (4.0.2): do not edit by hand
\name{alltimerecord}
\alias{alltimerecord}
\title{Function to List the all-time records of a team}
\usage{
alltimerecord(df, teamname)
}
\arguments{
\item{df}{df}

\item{teamname}{teamname}
}
\description{
Function to List the all-time records of a team
}
\examples{
df <- engsoccerdata2
alltimerecord(df, "Aston Villa")
alltimerecord(df, "Arsenal")
alltimerecord(df, "Liverpool")
alltimerecord(df, "Manchester United")
alltimerecord(df, "York City")
alltimerecord(df, "Rochdale")
alltimerecord(df, "Birmingham City")
alltimerecord(df, "Leeds City")
}

