% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/belgium_current.R
\name{belgium_current}
\alias{belgium_current}
\title{Get current Belgium season data for top tier}
\usage{
belgium_current()
}
\value{
a dataframe with results for current
season for top division
}
\description{
Get current Belgium season data for top tier
}
\examples{
belgium_current()
}
