% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bestwins}
\alias{bestwins}
\title{Return each team's record win (top 5 etc)}
\usage{
bestwins(df, teamname, type = NULL, N = NULL)
}
\description{
Return each team's record win (top 5 etc)
}
\examples{
bestwins(df,"Everton")
bestwins(df,"Aston Villa", type="H")
bestwins(df,"York City", type="A")
bestwins(df,"Port Vale", N=20)
bestwins(df,"Hull City", type="A", N=7)
}

