% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/engsoccerdata-package.R
\docType{data}
\name{bundesliga}
\alias{bundesliga}
\title{Top tier results for Germany (1963-2014)}
\format{\preformatted{'data.frame':	15508 obs. of  8 variables:
 $ date   : chr  "1963-08-24" "1963-08-24" "1963-08-24" "1963-08-24" ...
 $ Season : int  1963 1963 1963 1963 1963 1963 1963 1963 1963 1963 ...
 $ home   : chr  "Werder Bremen" "1. FC Saarbrücken" "TSV 1860 München" "Frankfurter SG Eintracht" ...
 $ visitor: chr  "Borussia Dortmund" "1. FC Köln" "Eintracht Braunschweig" "1. FC Kaiserslautern" ...
 $ FT     : chr  "3-2" "0-2" "1-1" "1-1" ...
 $ hgoal  : int  3 0 1 1 2 1 1 1 3 4 ...
 $ vgoal  : int  2 2 1 1 0 1 4 1 3 2 ...
 $ tier   : int  1 1 1 1 1 1 1 1 1 1 ...
}}
\usage{
bundesliga
}
\description{
Top tier results for Germany (1963-2014)
}
\keyword{datasets}

