% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/champs.R
\docType{data}
\name{champs}
\alias{champs}
\title{European Cup and Champions League Results 1955-2016}
\format{A data frame with 4866 rows and 11 variables:
\describe{
  \item{Date}{Date of match}
  \item{Season}{Season of match - refers to starting year}
  \item{round}{round}
  \item{leg}{leg}
  \item{home}{Home team}
  \item{visitor}{Visiting team}
  \item{FT}{Full-time result}
  \item{HT}{Half-time result}
  \item{aet}{After Extra Time result}
  \item{pens}{Penalties result}
  \item{tiewinner}{Eventual Tie Winner}
}}
\usage{
champs
}
\description{
All results for European Cup and Champions League
matches including qualifiers - NOTE: This is
still incomplete and in progress. Will be cleaned
soon.
}
\keyword{datasets}

