% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deductions.R
\docType{data}
\name{deductions}
\alias{deductions}
\title{Points Deductions}
\format{A data frame with 45 rows and 5 variables:
\describe{
  \item{Season}{Season of match - refers to starting year}
  \item{team}{Team}
  \item{tier}{Tier of football pyramid}
  \item{deduction}{Total points deducted from final table}
  \item{country}{Country of League}
}}
\usage{
deductions
}
\description{
Points deductions made to teams by season
}
\keyword{datasets}
