% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/england1939.R
\docType{data}
\name{england1939}
\alias{england1939}
\title{English league results 1939}
\format{A data frame with 190096 rows and 12 variables:
\describe{
  \item{Date}{Date of match}
  \item{Season}{Season of match - refers to starting year}
  \item{home}{Home team}
  \item{visitor}{Visiting team}
  \item{FT}{Full-time result}
  \item{hgoal}{Goals scored by home team}
  \item{vgoal}{Goals scored by visiting team}
  \item{division}{Division: 1,2,3,4 or 3a (Old 3-North) or 3b (Old 3-South)}
  \item{tier}{Tier of football pyramid: 1,2,3,4}
  \item{totgoal}{Total goals in game}
  \item{goaldif}{Goal difference in game home goals - visitor goals}
  \item{result}{Result: H-Home Win, A-Away Win, D-Draw}
}}
\usage{
england1939
}
\description{
All results for English soccer games the abandoned
1939/40 season.
}
\keyword{datasets}

