% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/england5_current.R
\name{england5_current}
\alias{england5_current}
\title{Get current England season data for tier 5}
\usage{
england5_current(Season = 2025)
}
\arguments{
\item{Season}{The current Season (starting year, e.g. 2025 for 2025/26)}
}
\value{
A dataframe with results for current season for tier 5 (National League)
}
\description{
Get current England season data for tier 5
}
