% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/england_club_data.R
\docType{data}
\name{england_club_data}
\alias{england_club_data}
\title{English club data}
\format{
A data frame with 660 rows and 10 variables:
\describe{
  \item{team}{the name of the team in the matches dataset}
  \item{lat}{the latitude of the teams ground (beta- drawn from the locations file)}
  \item{lon}{the longitude of the teams ground (beta- drawn from the locations file)}
  \item{highest_div}{the highest division (tier) the team reaches in the matches dataset}
  \item{col1}{primary team colour}
  \item{col2}{secondary team colour}
  \item{short_name}{an abbreviated name for the team in case of long names}
  \item{three_letter_name}{a 3 letter code for the team (not added)}
  \item{nicknames}{a list of nicknames for the team (not added- see team names file)}
  \item{extinct}{Whether club no longer exist or not}
}
}
\usage{
england_club_data
}
\description{
additional data for english clubs (locations, colours, names)
for teams in the engsoccerdata dataset
currently beta and needs filling in
}
\keyword{datasets}
