% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/england_current.R
\name{england_current}
\alias{england_current}
\title{Get current England season data for all tiers}
\usage{
england_current(Season = 2018)
}
\arguments{
\item{Season}{the current Season}
}
\value{
a dataframe with results for current
season for all top four divisions
}
\description{
Get current England season data for all tiers
}
\examples{
england_current()
}
