% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/englandplayoffs.R
\docType{data}
\name{englandplayoffs}
\alias{englandplayoffs}
\title{English league playoff results 1892-2016}
\format{A data frame with 496 rows and 18 variables:
\describe{
  \item{Date}{Date of match}
  \item{Season}{Season of match - refers to starting year}
  \item{home}{Home team}
  \item{visitor}{Visiting team}
  \item{FT}{Full-time result}
  \item{hgoal}{Goals scored by home team}
  \item{vgoal}{Goals scored by visiting team}
  \item{division}{Division for playoff}
  \item{round}{Round of playoffs}
  \item{tie}{Initial tie or replay}
  \item{htier}{Tier of football pyramid of home team}
  \item{vtier}{Tier of football pyramid of visiting team}
  \item{aet}{Did game go to extra time?}
  \item{pen}{Did game go to penalties?}
  \item{pens}{Result of penalties}
  \item{Venue}{Venue}
  \item{attendance}{Attendance}
  \item{neutral}{Was game played at neutal venue?}
}}
\usage{
englandplayoffs
}
\description{
All results for English league playoff games in the top tier
from 1892/93 season to 2015/16 season.
}
\keyword{datasets}

