% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/engsoccerdata-package.R
\docType{data}
\name{engsoccerdata2}
\alias{engsoccerdata2}
\title{Results and dates of all top 4 tier soccer games in England 1888-2015}
\format{\preformatted{'data.frame':	190096 obs. of  12 variables:
 $ Date    : chr  "1888-12-15" "1889-01-19" "1889-03-23" "1888-12-01" ...
 $ Season  : int  1888 1888 1888 1888 1888 1888 1888 1888 1888 1888 ...
 $ home    : chr  "Accrington F.C." "Accrington F.C." "Accrington F.C." "Accrington F.C." ...
 $ visitor : chr  "Aston Villa" "Blackburn Rovers" "Bolton Wanderers" "Burnley" ...
 $ FT      : chr  "1-1" "0-2" "2-3" "5-1" ...
 $ hgoal   : int  1 0 2 5 6 3 1 0 2 2 ...
 $ vgoal   : int  1 2 3 1 2 1 2 0 0 1 ...
 $ division: chr  "1" "1" "1" "1" ...
 $ tier    : int  1 1 1 1 1 1 1 1 1 1 ...
 $ totgoal : int  2 2 5 6 8 4 3 0 2 3 ...
 $ goaldif : int  0 -2 -1 4 4 2 -1 0 2 1 ...
 $ result  : chr  "D" "A" "A" "H" ...
}}
\usage{
engsoccerdata2
}
\description{
Results and dates of all top 4 tier soccer games in England 1888-2015
}
\keyword{datasets}

