% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{engsoccerteams}
\alias{engsoccerteams}
\title{List of 142 teams plus whether they were in the top 4 divisions in 2013/4}
\format{\preformatted{'data.frame':	142 obs. of  2 variables:
 $ team       : chr  "Aberdare Athletic" "Accrington" "Accrington F.C." "Accrington Stanley" ...
 $ current2013: chr  "no" "yes" "no" "no" ...
}}
\usage{
engsoccerteams
}
\description{
List of 142 teams plus whether they were in the top 4 divisions in 2013/4
}
\keyword{datasets}

