% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/engsoccerdata-package.R
\docType{data}
\name{facup}
\alias{facup}
\title{Results and dates of all FA Cup soccer games in England 1871-2014}
\format{\preformatted{'data.frame':	16323 obs. of  19 variables:
 $ date      : chr  "1871-11-11" "1871-11-11" "1871-11-11" "1871-11-11" ...
 $ Season    : int  1871 1871 1871 1871 1871 1871 1871 1871 1871 1871 ...
 $ home      : chr  "Barnes" "Hitchin" "Maidenhead United" "Upton Park" ...
 $ visitor   : chr  "Civil Service" "Crystal Palace 1861" "Marlow" "Clapham Rovers" ...
 $ FT        : chr  "2-0" "0-0" "2-0" "0-3" ...
 $ hgoal     : int  2 0 2 0 3 3 1 1 0 0 ...
 $ vgoal     : int  0 0 0 3 0 1 1 0 5 0 ...
 $ round     : chr  "1" "1" "1" "1" ...
 $ tie       : chr  "initial" "initial" "initial" "initial" ...
 $ aet       : chr  NA NA NA NA ...
 $ pen       : chr  NA NA NA NA ...
 $ pens      : chr  NA NA NA NA ...
 $ hp        : int  NA NA NA NA NA NA NA NA NA NA ...
 $ vp        : int  NA NA NA NA NA NA NA NA NA NA ...
 $ Venue     : chr  NA "Top Field, Hitchin" "York Road, Maidenhead" "West Ham Park, London" ...
 $ attendance: chr  "1,200" "750" "1,287" "1,500" ...
 $ nonmatch  : chr  NA NA NA NA ...
 $ notes     : chr  NA "No replay, both teams progress" NA NA ...
 $ neutral   : chr  NA NA NA NA ...
}}
\usage{
facup
}
\description{
Results and dates of all FA Cup soccer games in England 1871-2014
}
\keyword{datasets}

