% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/engsoccerdata-package.R
\docType{data}
\name{facupteams}
\alias{facupteams}
\title{List of 500+ teams who have played in the FA Cup proper with earliest and most recent years}
\format{\preformatted{'data.frame':	540 obs. of  3 variables:
 $ team: chr  "105th Regiment" "1st Surrey Rifles" "Aberdare Athletic" "Accrington" ...
 $ min : int  1875 1872 1922 1881 1906 1879 1925 2003 1993 2008 ...
 $ max : int  1878 1886 1927 1893 2013 1885 2013 2008 2011 2013 ...
}}
\usage{
facupteams
}
\description{
List of 500+ teams who have played in the FA Cup proper with earliest and most recent years
}
\keyword{datasets}

