% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/france.R
\docType{data}
\name{france}
\alias{france}
\title{French league results 1933-2017}
\format{A data frame with 26767 rows and 12 variables:
\describe{
  \item{Date}{Date of match}
  \item{Season}{Season of match - refers to starting year}
  \item{home}{Home team}
  \item{visitor}{Visiting team}
  \item{FT}{Full-time result}
  \item{hgoal}{Goals scored by home team}
  \item{vgoal}{Goals scored by visiting team}
  \item{division}{Division: 1}
  \item{tier}{Tier of football pyramid: 1}
  \item{totgoal}{Total goals in game}
  \item{goaldif}{Goal difference in game home goals - visitor goals}
  \item{result}{Result: H-Home Win, A-Away Win, D-Draw}
}}
\usage{
france
}
\description{
All results for French soccer games in the top tier
from 1933/34 season to 2016/17 season.  Doesn't include
playoff games. Note: this dataset is as is, it has not
been proof checked.
}
\keyword{datasets}
