% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/france_current.R
\name{france_current}
\alias{france_current}
\title{Get current France season data for top tier}
\usage{
france_current(Season = 2017)
}
\arguments{
\item{Season}{the current Season}
}
\value{
a dataframe with results for current
season for top division
}
\description{
Get current France season data for top tier
}
\examples{
france_current()
}
