% Generated by roxygen2 (4.0.2): do not edit by hand
\name{games_between.summary}
\alias{games_between.summary}
\title{Function to List the summary stats of all games ever between two teams}
\usage{
games_between.summary(df, teamname1, teamname2)
}
\arguments{
\item{df}{df}

\item{teamname1}{teamname1}

\item{teamname2}{teamname2}
}
\description{
Function to List the summary stats of all games ever between two teams
}
\examples{
df <- engsoccerdata2
games_between.summary(df, "Exeter City", "York City")
games_between.summary(df, "Aston Villa", "York City")
games_between.summary(df, "Manchester United", "Liverpool")
games_between.summary(df, "Aston Villa", "Everton")
games_between.summary(df, "Sheffield Wednesday", "Sheffield United")
games_between.summary(df, "Queens Park Rangers", "Fulham")
}

