% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_between_sum.R
\name{games_between_sum}
\alias{games_between_sum}
\title{Function to List the summary stats of all games ever between two teams}
\usage{
games_between_sum(df = NULL, teamname1 = NULL, teamname2 = NULL)
}
\arguments{
\item{df}{The results dataframe}

\item{teamname1}{teamname1}

\item{teamname2}{teamname2}
}
\value{
a dataframe with summary of results. Will be
empty dataframe if never played.
}
\description{
Function to List the summary stats of all games ever between two teams
}
\examples{
games_between_sum(england, "Exeter City", "York City")
games_between_sum(england, "Aston Villa", "York City")
games_between_sum(england, "Manchester United", "Liverpool")
}

