% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/games_by_tier.R
\name{games_by_tier}
\alias{games_by_tier}
\title{Number of Games in Each Tier}
\usage{
games_by_tier(df, Tier = NULL)
}
\arguments{
\item{df}{df}

\item{Tier}{Tier}
}
\description{
Number of Games in Each Tier
}
\examples{
df <- engsoccerdata2
games_by_tier(df, 1)
games_by_tier(df, 2)
games_by_tier(df, 3)
games_by_tier(df, 4)

as.data.frame(unclass(games_by_tier(df))) #all games by all teams in all tiers
}

