% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germany_current.R
\name{germany_current}
\alias{germany_current}
\title{Get current Germany season data for top tiers}
\usage{
germany_current(Season = 2025)
}
\arguments{
\item{Season}{the current Season}
}
\value{
a dataframe with results for current season for Bundesliga (tier 1)
and 2. Bundesliga (tier 2)
}
\description{
Get current Germany season data for top tiers
}
\examples{
germany_current()
}
