% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCurrentData.R
\name{getCurrentData}
\alias{getCurrentData}
\title{Get current season data}
\usage{
getCurrentData(df, division, tier, Season)
}
\arguments{
\item{df}{the results dataset}

\item{division}{division to add}

\item{tier}{tier to add}

\item{Season}{Season to get current data for}
}
\value{
a dataframe with results for current season
}
\description{
Get current season data
}
