% Generated by roxygen2 (4.0.2): do not edit by hand
\name{goals_by_team}
\alias{goals_by_team}
\title{Return all instances of a team scoring n goals}
\usage{
goals_by_team(df, goals, teamname)
}
\description{
Return all instances of a team scoring n goals
}
\examples{
df <- engsoccerdata2
goals_by_team(df, 7, "York City")
goals_by_team(df, 8, "Manchester United")
goals_by_team(df, 9, "Aston Villa")
}

