% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greece_current.R
\name{greece_current}
\alias{greece_current}
\title{Get current Greece season data for top tier}
\usage{
greece_current(Season = 2017)
}
\arguments{
\item{Season}{the current Season}
}
\value{
a dataframe with results for current
season for top division
}
\description{
Get current Greece season data for top tier
}
\examples{
greece_current()
}
