% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holland.R
\docType{data}
\name{holland}
\alias{holland}
\title{Dutch league results 1956-2016}
\format{A data frame with 18096 rows and 8 variables:
\describe{
  \item{Date}{Date of match}
  \item{Season}{Season of match - refers to starting year}
  \item{home}{Home team}
  \item{visitor}{Visiting team}
  \item{FT}{Full-time result}
  \item{hgoal}{Goals scored by home team}
  \item{vgoal}{Goals scored by visiting team}
  \item{tier}{Tier of football pyramid: 1}
}}
\usage{
holland
}
\description{
All results for Dutch soccer games in the top tier
from 1956/57 season to 2015/16 season.
}
\keyword{datasets}

