% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/engsoccerdata-package.R
\docType{data}
\name{holland1}
\alias{holland1}
\title{Top tier results for Holland (1956-2014)}
\format{\preformatted{'data.frame':	17484 obs. of  8 variables:
 $ date   : chr  "1956-09-02" "1956-09-02" "1956-09-02" "1956-09-02" ...
 $ Season : int  1956 1956 1956 1956 1956 1956 1956 1956 1956 1956 ...
 $ home   : chr  "AFC Ajax" "BVV" "DOS Utrecht" "Fortuna Sittard" ...
 $ visitor: chr  "NAC Breda" "USV Elinkwijk" "Sparta Rotterdam" "SC Eindhoven" ...
 $ FT     : chr  "1-0" "1-2" "2-3" "4-1" ...
 $ hgoal  : int  1 1 2 4 1 1 5 1 3 0 ...
 $ vgoal  : int  0 2 3 1 3 3 2 0 3 6 ...
 $ tier   : int  1 1 1 1 1 1 1 1 1 1 ...
}}
\usage{
holland1
}
\description{
Top tier results for Holland (1956-2014)
}
\keyword{datasets}

