% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holland_current.R
\name{holland_current}
\alias{holland_current}
\title{Get current Holland season data for top tier}
\usage{
holland_current(Season = 2017)
}
\arguments{
\item{Season}{the current Season}
}
\value{
a dataframe with results for current
season for top division
}
\description{
Get current Holland season data for top tier
}
\examples{
holland_current()
}
