% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homeaway.R
\name{homeaway}
\alias{homeaway}
\title{Return all results by team}
\usage{
homeaway(df)
}
\arguments{
\item{df}{the results dataset}
}
\value{
a dataframe of all games by team
}
\description{
Return all results by team
}
\examples{
homeaway(england[england$Season==2015 & england$tier==1,])

}
