% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/italy.R
\docType{data}
\name{italy}
\alias{italy}
\title{Italy league results 1934-2016}
\format{A data frame with 25404 rows and 8 variables:
\describe{
  \item{Date}{Date of match}
  \item{Season}{Season of match - refers to starting year}
  \item{home}{Home team}
  \item{visitor}{Visiting team}
  \item{FT}{Full-time result}
  \item{hgoal}{Goals scored by home team}
  \item{vgoal}{Goals scored by visiting team}
  \item{tier}{Tier of football pyramid: 1}
}}
\usage{
italy
}
\description{
All results for Italian soccer games in the top tier
from 1934/35 season to 2015/16 season.
}
\keyword{datasets}

