% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/italy_current.R
\name{italy_current}
\alias{italy_current}
\title{Get current Italy season data for top tier}
\usage{
italy_current(Season = 2017)
}
\arguments{
\item{Season}{the current Season}
}
\value{
a dataframe with results for current
season for top division
}
\description{
Get current Italy season data for top tier
}
\examples{
italy_current()
}
