% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/engsoccerdata-package.R
\docType{data}
\name{italycalcio}
\alias{italycalcio}
\title{Top tier results for Italy (1934-2014)}
\format{\preformatted{'data.frame':	24644 obs. of  8 variables:
 $ date   : chr  "1934-09-30" "1934-09-30" "1934-09-30" "1934-09-30" ...
 $ Season : int  1934 1934 1934 1934 1934 1934 1934 1934 1934 1934 ...
 $ home   : chr  "Lazio Roma" "Torino FC" "Sampierdarenese" "SSC Napoli" ...
 $ visitor: chr  "US Livorno" "Unione Triestina" "Bologna FC" "US Alessandria 1912" ...
 $ FT     : chr  "6-1" "3-1" "2-1" "0-1" ...
 $ hgoal  : int  6 3 2 0 4 0 3 1 1 1 ...
 $ vgoal  : int  1 1 1 1 1 2 0 2 1 1 ...
 $ tier   : int  1 1 1 1 1 1 1 1 1 1 ...
}}
\usage{
italycalcio
}
\description{
Top tier results for Italy (1934-2014)
}
\keyword{datasets}

