% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaguecup.R
\docType{data}
\name{leaguecup}
\alias{leaguecup}
\title{English league playoff results 1892-2017}
\format{A data frame with 16871 rows and 19 variables:
\describe{
  \item{Date}{Date of match}
  \item{Season}{Season of match - refers to starting year}
  \item{home}{Home team}
  \item{visitor}{Visiting team}
  \item{FT}{Full-time result}
  \item{hgoal}{Goals scored by home team}
  \item{vgoal}{Goals scored by visiting team}
  \item{round}{Proper Round}
  \item{tie}{Initial tie or replay}
  \item{leg}{Which leg of a two legged match}
  \item{aet}{Did game go to extra time?}
  \item{pen}{Did game go to penalties?}
  \item{pens}{Result of penalties}
  \item{Venue}{Venue}
  \item{attendance}{Attendance}
  \item{northsouth}{}
}}
\usage{
leaguecup
}
\description{
All results for English league playoff games in the top tier
from 1892/93 season to 2016/17 season.
}
\keyword{datasets}
