% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maketable_all.R
\name{maketable_all}
\alias{maketable_all}
\title{Make a league table - not limited by seasons or tiers}
\usage{
maketable_all(df = NULL, Season = NULL, tier = NULL, pts = 3,
  begin = NULL, end = NULL, type = c("both", "home", "away"))
}
\arguments{
\item{df}{The results dataframe}

\item{Season}{The Season}

\item{tier}{The tier}

\item{pts}{Points for a win. Default is 3.}

\item{begin}{Earliest date of results to make table from (format Y-m-d)}

\item{end}{Latest date of results to make table from (format Y-m-d)}

\item{type}{Whether to show all results together or only home or away results.}
}
\value{
a dataframe with a league table
}
\description{
Make a league table - not limited by seasons or tiers
}
\section{Notes}{

The table that is produced is based upon 3 points for a win (unless otherwise
    defined), 1 for a draw and 0 for a loss.  The table is sorted based upon descending
    GD and then descending GF as tie-breakers. Use other 'maketable' functions for
    more precise tables for each league.
}

\examples{
maketable_all(df=england[england$tier==1,],begin="1992-08-15",
     end="2017-07-01") #EPL historical table
maketable_all(df=england[england$tier==1,],begin="1992-08-15",
     end="2017-07-01", type="away") #EPL historical table away results
}
