% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maketable_eng.R
\name{maketable_eng}
\alias{maketable_eng}
\title{Make an English league table}
\usage{
maketable_eng(df = NULL, Season = NULL, tier = NULL,
  division = NULL, penalties = FALSE)
}
\arguments{
\item{df}{The results dataframe}

\item{Season}{The Season}

\item{tier}{The tier - must be included}

\item{division}{The division}

\item{penalties}{Whether to account for points penalties}
}
\value{
a dataframe with a league table
}
\description{
Make an English league table
}
\section{Notes}{

This table makes league tables according to the points and tie-breaking procedures
that were in place for each league in each year
}

\examples{
maketable_eng(df=england,Season=1920,tier=1)
maketable_eng(df=england,Season=1947,division="3a",tier=3)
maketable_eng(df=england,Season=1975,tier=1)
maketable_eng(df=england,Season=1975,tier=2)
maketable_eng(df=england,Season=2007,tier=3, penalties=TRUE)
}
