% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/namecheck.R
\name{namecheck}
\alias{namecheck}
\title{Function to check what characters exist in which team names}
\usage{
namecheck(df, x)
}
\arguments{
\item{df}{df}

\item{x}{x}
}
\description{
Function to check what characters exist in which team names
}
\examples{
df <- engsoccerdata2
namecheck(df, "hester")
namecheck(df, "United")
namecheck(df, "Athletic")
namecheck(df, "Rovers")
namecheck(df, "x")
}

