% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngoals.R
\name{ngoals}
\alias{ngoals}
\title{Return all instances of a team scoring n goals}
\usage{
ngoals(df = NULL, n = NULL, teamname = NULL)
}
\arguments{
\item{df}{The results dataframe}

\item{n}{Number of goals}

\item{teamname}{teamname}
}
\value{
a dataframe with summary of results.
}
\description{
Return all instances of a team scoring n goals
}
\examples{
ngoals(england, 7, "York City")
ngoals(england, 8, "Manchester United")
ngoals(england, 9, "Aston Villa")

}

