% Generated by roxygen2 (4.0.2): do not edit by hand
\name{opponentfreq}
\alias{opponentfreq}
\title{function to see number of times  played against each opponent}
\usage{
opponentfreq(df, teamname)
}
\description{
function to see number of times  played against each opponent
}
\examples{
opponentfreq(df, "Exeter City")  #just gives top few opponents
opponentfreq(df, "Aston Villa")
opponentfreq(df, "Everton")
opponentfreq(df, "York City")
opponentfreq(df, "Bolton Wanderers")
opponentfreq(df, "Milton Keynes Dons")

as.data.frame(unclass(opponentfreq(df, "Exeter City"))) #view all opponents
}

