% Generated by roxygen2 (4.0.2): do not edit by hand
\name{opponents}
\alias{opponents}
\title{Total number of different opponents}
\usage{
opponents(df, Tier = NULL)
}
\arguments{
\item{df}{df}

\item{Tier}{Tier}
}
\description{
Total number of different opponents
}
\examples{
df <- engsoccerdata2
opponents(df)  #Grimsby Town and Lincoln City have the most unique opponents - 134 each
opponents(df,1)
opponents(df,2)
opponents(df,3)
opponents(df,4)
}

