% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/engsoccerdata-package.R
\docType{data}
\name{playoffs}
\alias{playoffs}
\title{All playoff matches including 1890s 'test-matches' and modern playoffs}
\format{\preformatted{'data.frame':	466 obs. of  18 variables:
 $ date      : chr  "1893-04-22" "1893-04-22" "1893-04-22" "1893-04-27" ...
 $ Season    : int  1892 1892 1892 1892 1893 1893 1893 1894 1894 1894 ...
 $ home      : chr  "Sheffield United" "Birmingham City" "Darwen" "Manchester United" ...
 $ visitor   : chr  "Accrington F.C." "Manchester United" "Notts County" "Birmingham City" ...
 $ FT        : chr  "1-0" "1-1" "3-2" "5-2" ...
 $ hgoal     : int  1 1 3 5 3 2 4 1 2 3 ...
 $ vgoal     : int  0 1 2 2 1 0 0 0 1 0 ...
 $ division  : chr  "test match" "test match" "test match" "test match" ...
 $ round     : chr  "f" "f" "f" "f" ...
 $ tie       : chr  "initial" "initial" "initial" "replay" ...
 $ htier     : int  2 2 2 1 2 2 1 2 1 1 ...
 $ vtier     : int  1 1 1 2 1 1 2 1 2 2 ...
 $ aet       : chr  NA NA NA NA ...
 $ pen       : chr  NA NA NA NA ...
 $ pens      : chr  NA NA NA NA ...
 $ Venue     : chr  "Castle Ground, Nottingham" "Victoria Ground, Stoke" "Hyde Road, Manchester" "Olive Grove, Sheffield" ...
 $ attendance: chr  "6,000" "4,000" "5,000" "6,000" ...
 $ neutral   : chr  "yes" "yes" "yes" "yes" ...
}}
\usage{
playoffs
}
\description{
All playoff matches including 1890s 'test-matches' and modern playoffs
}
\keyword{datasets}

