% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portugal.R
\docType{data}
\name{portugal}
\alias{portugal}
\title{Portuguese league results 1994-2017}
\format{A data frame with 6510 rows and 12 variables:
\describe{
  \item{Date}{Date of match}
  \item{Season}{Season of match - refers to starting year}
  \item{home}{Home team}
  \item{visitor}{Visiting team}
  \item{FT}{Full-time result}
  \item{hgoal}{Goals scored by home team}
  \item{vgoal}{Goals scored by visiting team}
  \item{division}{Division}
  \item{tier}{Tier of football pyramid: 1}
  \item{totgoal}{Total goals in game}
  \item{goaldif}{Goal difference in game home goals - visitor goals}
  \item{result}{Result: H-Home Win, A-Away Win, D-Draw}
}}
\usage{
portugal
}
\description{
All results for Portuguese soccer games in the top tier
from 1994/95 season to 2016/17 season.  Doesn't include
playoff games.
}
\keyword{datasets}
