% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portugal_current.R
\name{portugal_current}
\alias{portugal_current}
\title{Get current Portugal season data for top tier}
\usage{
portugal_current(Season = 2017)
}
\arguments{
\item{Season}{the current Season}
}
\value{
a dataframe with results for current
season for top division
}
\description{
Get current Portugal season data for top tier
}
\examples{
portugal_current()
}
