% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/score_most.R
\name{score_most}
\alias{score_most}
\title{This returns the team who has been involved in the most games of each scoreline}
\usage{
score_most(df, score)
}
\arguments{
\item{df}{df}

\item{score}{score}
}
\description{
This returns the team who has been involved in the most games of each scoreline
}
\examples{
df <- engsoccerdata2
score_most(df, "6-6") # Arsenal 1  Charlton Athletic 1  Leicester City 1  Middlesbrough 1
score_most(df, "5-5") # Blackburn Rovers 3   West Ham United 3
score_most(df, "4-4") # Tottenham Hotspur 14
score_most(df, "3-3") # Manchester City 68   Wolverhampton Wanderers 68
score_most(df, "2-2") # Leicester City 274
score_most(df, "1-1") # Sheffield United 560
score_most(df, "0-0") # Notts County 363

score_most(df, "1-0") # Birmingham City 795 - most involved in 1-0 or 0-1 games
score_most(df, "8-0") # Arsenal 7 - most involved in 8-0 or 0-8 games
score_most(df, "9-1") # Notts County 4 - most involved in 4-1 or 1-4 games
}

