% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/score_team_all.R
\name{score_team.all}
\alias{score_team.all}
\title{Lists all matches that a team has played in that ended in a scoreline}
\usage{
score_team.all(df, score, teamname)
}
\arguments{
\item{df}{df}

\item{score}{score}

\item{teamname}{teamname}
}
\description{
Lists all matches that a team has played in that ended in a scoreline
}
\examples{
df <- engsoccerdata2
score_team.all(df, "1-0", "Arsenal")  # All games ending 1-0 or 0-1 involving Arsenal
score_team.all(df, "0-1", "Arsenal")  # All games ending 1-0 or 0-1 involving Arsenal
score_team.all(df, "4-3", "Coventry City")  # All games ending 4-3 or 3-4 involving Coventry City
score_team.all(df, "4-4", "Leeds United")  # All games ending 4-4 involving Leeds United
}

