% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_teamX.R
\name{score_teamX}
\alias{score_teamX}
\title{List all occurrences of a specific scoreline for a specific team}
\usage{
score_teamX(df = NULL, score = NULL, teamname = NULL)
}
\arguments{
\item{df}{the results dataset}

\item{score}{the scoreline}

\item{teamname}{the team}
}
\value{
a dataframe of games ending in that result
}
\description{
List all occurrences of a specific scoreline for a specific team
}
\examples{
score_teamX(england,"4-4", "Tottenham Hotspur")
#all 4-4 draws Tottenham Hotspur have played in (home and away)

score_teamX(england,"3-5", "York City")
#list all 5-3 defeats suffered by York City (regardless of if occurred home/away)

score_teamX(england,"5-3", "York City")
#list all 5-3 victories by York City (regardless of if occurred home/away)

score_teamX(england,"8-0", "Arsenal")
#list all 8-0 victories by Arsenal (regardless of if occurred home/away)

score_teamX(england,"0-8", "Arsenal")
#list all 8-0 defeats suffered by Arsenal (regardless of if occurred home/away)
}

