% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scoreline_by_team}
\alias{scoreline_by_team}
\title{How often each team has a won,lost,drawn by a scoreline ?}
\usage{
scoreline_by_team(df, score)
}
\description{
Note - teams with 0 occurrences will not be listed
}
\examples{
scoreline_by_team(df,"8-0")  #number of times each team has won 8-0 (regardless of if occurred home/away)
scoreline_by_team(df,"0-8")  #number of times each team has lost 8-0 (regardless of if occurred home/away)
scoreline_by_team(df, "6-6") #number of times each team has drawn 6-6 (regardless of if occurred home/away)
scoreline_by_team(df, "7-5") #number of times each team has won 7-5 (regardless of if occurred home/away)
scoreline_by_team(df, "1-0") #number of times each team has won 1-0 (regardless of if occurred home/away)
}

