% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/engsoccerdata-package.R
\docType{data}
\name{spainliga}
\alias{spainliga}
\title{All La Liga matches 1929-2014}
\format{\preformatted{'data.frame':	23155 obs. of  12 variables:
 $ date   : chr  "1929-02-10" "1929-02-10" "1929-02-10" "1929-02-10" ...
 $ Season : int  1928 1928 1928 1928 1928 1928 1928 1928 1928 1928 ...
 $ home   : chr  "Arenas de Getxo" "Espanyol Barcelona" "Real Madrid" "Real Sociedad" ...
 $ visitor: chr  "Atlético Madrid" "Real Unión" "CE Europa" "Athletic Bilbao" ...
 $ HT     : chr  "0-2" "1-0" "0-0" "1-1" ...
 $ FT     : chr  "2-3" "3-2" "5-0" "1-1" ...
 $ hgoal  : int  2 3 5 1 0 1 9 0 3 5 ...
 $ vgoal  : int  3 2 0 1 2 2 0 3 1 2 ...
 $ tier   : int  1 1 1 1 1 1 1 1 1 1 ...
 $ round  : chr  "league" "league" "league" "league" ...
 $ group  : chr  NA NA NA NA ...
 $ notes  : chr  NA NA NA NA ...
}}
\usage{
spainliga
}
\description{
All La Liga matches 1929-2014
}
\keyword{datasets}

