% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/totalgoals_by_team.R
\name{totalgoals_by_team}
\alias{totalgoals_by_team}
\title{Return all instances of a team being involved in a game with n goals}
\usage{
totalgoals_by_team(df, goals, teamname)
}
\arguments{
\item{df}{df}

\item{goals}{goals}

\item{teamname}{teamname}
}
\description{
Return all instances of a team being involved in a game with n goals
}
\examples{
df <- engsoccerdata2
totalgoals_by_team(df, 10, "York City")
totalgoals_by_team(df, 10, "Manchester United")
totalgoals_by_team(df, 12, "Aston Villa")
}

