% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totgoals.R
\name{totgoals}
\alias{totgoals}
\title{Return all instances of a team being involved in a game with n goals}
\usage{
totgoals(df = NULL, N = NULL, teamname = NULL)
}
\arguments{
\item{df}{The results dataframe}

\item{N}{Number of goals}

\item{teamname}{teamname}
}
\value{
a dataframe with  matches
}
\description{
Return all instances of a team being involved in a game with n goals
}
\examples{
totgoals(england, 10, "York City")
totgoals(england, 12, "Aston Villa")

}

