% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turkey_current.R
\name{turkey_current}
\alias{turkey_current}
\title{Get current Turkey season data for top tier}
\usage{
turkey_current(Season = 2017)
}
\arguments{
\item{Season}{the current Season}
}
\value{
a dataframe with results for current
season for top division
}
\description{
Get current Turkey season data for top tier
}
\examples{
turkey_current()
}
