% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alltimerecord.R
\name{alltimerecord}
\alias{alltimerecord}
\title{Get a team's all-time record}
\usage{
alltimerecord(df = NULL, teamname = NULL)
}
\arguments{
\item{df}{the results dataset}

\item{teamname}{team name}
}
\value{
a dataframe of all-time records
}
\description{
Get a team's all-time record
}
\examples{
alltimerecord(england, "Aston Villa")
alltimerecord(england, "York City")
alltimerecord(england, "Rochdale")
}

