% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bestwins.R
\name{bestwins}
\alias{bestwins}
\title{Return each team's record win (top 5 etc)}
\usage{
bestwins(df, teamname, type = NULL, N = NULL)
}
\arguments{
\item{df}{df}

\item{teamname}{teamname}

\item{type}{type}

\item{N}{N}
}
\description{
Return each team's record win (top 5 etc)
}
\examples{
df <- engsoccerdata2
bestwins(df,"Everton")
bestwins(df,"Aston Villa", type="H")
bestwins(df,"York City", type="A")
bestwins(df,"Port Vale", N=20)
bestwins(df,"Hull City", type="A", N=7)
}

