% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestwins.R
\name{bestwins}
\alias{bestwins}
\title{Return each team's best wins}
\usage{
bestwins(df = NULL, teamname = NULL, type = NULL, N = NULL)
}
\arguments{
\item{df}{the results dataset}

\item{teamname}{team name}

\item{type}{If \code{=NULL} then all results are returned. If
Otherwise valid types are \code{H}, \code{A}
relating to home-losses, away-losses}

\item{N}{The total number of games to return}
}
\value{
a dataframe of games ending in best wins
}
\description{
Return each team's best wins
}
\examples{
bestwins(england,"Everton")
bestwins(england,"Aston Villa", type="H")
bestwins(england,"York City", type="A")
bestwins(england,"Port Vale", N=20)

}

