% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/england5_current.R
\name{england5_current}
\alias{england5_current}
\title{Get current Enland season data for tier 5}
\usage{
england5_current(Season = 2017)
}
\arguments{
\item{Season}{the current Season}
}
\value{
a dataframe with results for current
season for top two divisions
}
\description{
Get current Enland season data for tier 5
}
\examples{
england5_current()
}
