% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facup.R
\docType{data}
\name{facup}
\alias{facup}
\title{English league playoff results 1892-2016}
\format{A data frame with 16625 rows and 19 variables:
\describe{
  \item{Date}{Date of match}
  \item{Season}{Season of match - refers to starting year}
  \item{home}{Home team}
  \item{visitor}{Visiting team}
  \item{FT}{Full-time result}
  \item{hgoal}{Goals scored by home team}
  \item{vgoal}{Goals scored by visiting team}
  \item{round}{Proper Round}
  \item{tie}{Initial tie or replay}
  \item{aet}{Did game go to extra time?}
  \item{pen}{Did game go to penalties?}
  \item{pens}{Result of penalties}
  \item{hp}{Total home penalties scored}
  \item{vp}{Total visitor penalties scored}
  \item{Venue}{Venue}
  \item{attendance}{Attendance}
  \item{nonmatch}{Was game nonmatch?}
  \item{notes}{Notes}
  \item{neutral}{Was game played at neutal venue?}
}}
\usage{
facup
}
\description{
All results for English league playoff games in the top tier
from 1892/93 season to 2015/16 season.
}
\keyword{datasets}

