% Generated by roxygen2 (4.0.2): do not edit by hand
\name{games_between}
\alias{games_between}
\title{Function to List all games ever between two teams}
\usage{
games_between(df, teamname1, teamname2)
}
\arguments{
\item{df}{df}

\item{teamname1}{teamname1}

\item{teamname2}{teamname2}
}
\description{
Function to List all games ever between two teams
}
\examples{
df <- engsoccerdata2
games_between(df, "Aston Villa", "York City")
games_between(df, "Carlisle United", "Chelsea")
games_between(df, "Manchester United", "Milton Keynes Dons") #will give 0 rows as never played
}

