% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_between.R
\name{games_between}
\alias{games_between}
\title{Function to List all games ever between two teams}
\usage{
games_between(df = NULL, teamname1 = NULL, teamname2 = NULL)
}
\arguments{
\item{df}{The results dataframe}

\item{teamname1}{teamname1}

\item{teamname2}{teamname2}
}
\value{
a dataframe of games between the two teams. If
they have never played will return an empty dataframe.
}
\description{
Function to List all games ever between two teams
}
\examples{
games_between(england, "Aston Villa", "York City")
games_between(england, "Carlisle United", "Chelsea")

}

