% Generated by roxygen2 (4.0.2): do not edit by hand
\name{games_by_tier}
\alias{games_by_tier}
\title{Number of Games in Each Tier}
\usage{
games_by_tier(df, Tier = NULL)
}
\description{
Number of Games in Each Tier
}
\examples{
df <- engsoccerdata2
games_by_tier(df, 1)
games_by_tier(df, 2)
games_by_tier(df, 3)
games_by_tier(df, 4)

as.data.frame(unclass(games_by_tier(df))) #all games by all teams in all tiers
}

