% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germany_current.R
\name{germany_current}
\alias{germany_current}
\title{Get current Germany season data for all tiers}
\usage{
germany_current(Season = 2017)
}
\arguments{
\item{Season}{the current Season}
}
\value{
a dataframe with results for current
season for top two divisions
}
\description{
Get current Germany season data for all tiers
}
\examples{
germany_current()
}
