% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/goals_by_team.R
\name{goals_by_team}
\alias{goals_by_team}
\title{Return all instances of a team scoring n goals}
\usage{
goals_by_team(df, goals, teamname)
}
\arguments{
\item{df}{df}

\item{goals}{goals}

\item{teamname}{teamname}
}
\description{
Return all instances of a team scoring n goals
}
\examples{
df <- engsoccerdata2
goals_by_team(df, 7, "York City")
goals_by_team(df, 8, "Manchester United")
goals_by_team(df, 9, "Aston Villa")
}

