% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maketable.R
\name{maketable}
\alias{maketable}
\title{Make a league table}
\usage{
maketable(df = NULL, Season = NULL, tier = NULL, pts = 3)
}
\arguments{
\item{df}{The results dataframe}

\item{Season}{The Season}

\item{tier}{The tier}

\item{pts}{Points for a win. Default is 3.}
}
\value{
a dataframe with a league table
}
\description{
Make a league table
}
\section{Notes}{

The table that is produced is based upon 3 points for a win (unless otherwise
defined), 1 for a draw and 0 for a loss.  The table is sorted based upon descending
GD and then descending GF as tie-breakers. Different leagues have had different
methods for tie-breaks over the years. This league table is a simple generic one.
It also does not evaluate points deducted from teams or if games were  artificially
awarded to one side based on games not being played.
}
\examples{
maketable(df=england,Season=2013,tier=1,pts=3)
}

