% Generated by roxygen2 (4.0.2): do not edit by hand
\name{namecheck}
\alias{namecheck}
\title{Function to check what characters exist in which team names}
\usage{
namecheck(df, x)
}
\description{
Function to check what characters exist in which team names
}
\examples{
df <- engsoccerdata2
namecheck(df, "hester")
namecheck(df, "United")
namecheck(df, "Athletic")
namecheck(df, "Rovers")
namecheck(df, "x")
}

