% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opponentfreq.r
\name{opponentfreq}
\alias{opponentfreq}
\title{The number of times a team has played against each opponent}
\usage{
opponentfreq(df = NULL, teamname = NULL)
}
\arguments{
\item{df}{The results dataframe}

\item{teamname}{teamname}
}
\value{
a dataframe with frequency of matches and team.
}
\description{
The number of times a team has played against each opponent
}
\examples{
opponentfreq(england, "Aston Villa")
opponentfreq(england, "York City")
opponentfreq(england, "Milton Keynes Dons")
}

