% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opponents.R
\name{opponents}
\alias{opponents}
\title{Total number of unique opponents}
\usage{
opponents(df = NULL, Tier = NULL)
}
\arguments{
\item{df}{A results dataframe}

\item{Tier}{Tier}
}
\value{
a dataframe with teams and frequency of unique opponents
}
\description{
Total number of unique opponents
}
\examples{
opponents(england)
opponents(england,4)

}

