% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_most.R
\name{score_most}
\alias{score_most}
\title{Team who has been involved in the most games of each scoreline}
\usage{
score_most(df, score)
}
\arguments{
\item{df}{Results dataframe}

\item{score}{score}
}
\value{
a dataframe with frequency of matches and team.
}
\description{
Team who has been involved in the most games of each scoreline
}
\examples{
score_most(england, "6-6")
score_most(england, "8-0")
score_most(england, "9-1")
}

