% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_team.R
\name{score_team}
\alias{score_team}
\title{Lists all matches that a team has played in that ended in a scoreline}
\usage{
score_team(df = NULL, score = NULL, teamname = NULL)
}
\arguments{
\item{df}{the results dataset}

\item{score}{the scoreline}

\item{teamname}{the team}
}
\value{
a dataframe of games ending in that result
}
\description{
Lists all matches that a team has played in that ended in a scoreline
}
\examples{
score_team(england, "1-0", "Arsenal")  # All games ending 1-0 or 0-1 involving Arsenal
score_team(england, "4-3", "Coventry City")  # All games ending 4-3 or 3-4 involving Coventry City
}

