% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scotland_current.R
\name{scotland_current}
\alias{scotland_current}
\title{Get current Scotland season data for all tiers}
\usage{
scotland_current(Season = 2017)
}
\arguments{
\item{Season}{the current Season}
}
\value{
a dataframe with results for current
season for top four divisions
}
\description{
Get current Scotland season data for all tiers
}
\examples{
scotland_current()
}
