% Generated by roxygen2 (4.0.2): do not edit by hand
\name{seasons_by_tier}
\alias{seasons_by_tier}
\title{Number of Seasons in Each Tier}
\usage{
seasons_by_tier(df, Tier = NULL)
}
\arguments{
\item{df}{df}

\item{Tier}{Tier}
}
\description{
Number of Seasons in Each Tier
}
\examples{
df <- engsoccerdata2
seasons_by_tier(df, 1)
seasons_by_tier(df, 2)
seasons_by_tier(df, 3)
seasons_by_tier(df, 4)

as.data.frame(unclass(seasons_by_tier(df))) #all seasons by all teams in all tiers
}

