% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spain_current.R
\name{spain_current}
\alias{spain_current}
\title{Get current Spain season data for top tier}
\usage{
spain_current(Season = 2017)
}
\arguments{
\item{Season}{the current Season}
}
\value{
a dataframe with results for current
season for top division
}
\description{
Get current Spain season data for top tier
}
\examples{
spain_current()
}
