% Generated by roxygen2 (4.0.2): do not edit by hand
\name{totalgoals_by_team}
\alias{totalgoals_by_team}
\title{Return all instances of a team being involved in a game with n goals}
\usage{
totalgoals_by_team(df, goals, teamname)
}
\description{
Return all instances of a team being involved in a game with n goals
}
\examples{
totalgoals_by_team(df, 10, "York City")
totalgoals_by_team(df, 10, "Manchester United")
totalgoals_by_team(df, 12, "Aston Villa")
}

