% Generated by roxygen2 (4.0.2): do not edit by hand
\name{worstlosses}
\alias{worstlosses}
\title{Return each team's worst losses}
\usage{
worstlosses(df, teamname, type = NULL, N = NULL)
}
\description{
Return each team's worst losses
}
\examples{
df <- engsoccerdata2
worstlosses(df,"Everton")
worstlosses(df,"Aston Villa", type="H")
worstlosses(df,"York City", type="A")
worstlosses(df,"Port Vale", N=20)
worstlosses(df,"Hull City", type="A", N=7)
}

